backend/fastapi/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
stripe==11.5.0
supabase==2.4.0
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
